+++ Exec/Execute_freedom.c
@@ -1,16 +1,5 @@
-/* ************************************************************************** */
-/*                                                                            */
-/*                                                        :::      ::::::::   */
-/*   Freedom_errors.c                                   :+:      :+:    :+:   */
-/*                                                    +:+ +:+         +:+     */
-/*   By: astefane <astefane@student.42madrid.com    +#+  +:+       +#+        */
-/*                                                +#+#+#+#+#+   +#+           */
-/*   Created: 2025/04/12 17:31:08 by astefane          #+#    #+#             */
-/*   Updated: 2025/05/06 16:08:58 by astefane         ###   ########.fr       */
-/*                                                                            */
-/* ************************************************************************** */
 
-#include "../libft.h"
+#include "../Mini.h"
 
 void	free_struct(t_pipex *data, char *message, int exit_code, int std)
 {
