+++ Exec/Execute_heredoc.c
@@ -1,16 +1,5 @@
-/* ************************************************************************** */
-/*                                                                            */
-/*                                                        :::      ::::::::   */
-/*   heredoc.c                                          :+:      :+:    :+:   */
-/*                                                    +:+ +:+         +:+     */
-/*   By: astefane <astefane@student.42madrid.com    +#+  +:+       +#+        */
-/*                                                +#+#+#+#+#+   +#+           */
-/*   Created: 2025/04/30 19:56:24 by astefane          #+#    #+#             */
-/*   Updated: 2025/05/08 18:51:41 by astefane         ###   ########.fr       */
-/*                                                                            */
-/* ************************************************************************** */
 
-#include "../libft.h"
+#include "../Mini.h"
 
 int	is_limiter(char *line, char *limiter)
 {
