+++ Exec/Execute_parse.c
@@ -1,4 +1,5 @@
-#include "../libft.h"
+
+#include "../Mini.h"
 
 t_command	*parse_commands(t_token *token, t_pipex *data)
 {
