+++ Varibles/variables.c
@@ -1,5 +1,5 @@
 
-// #include "../Mini.h"
+#include "../Mini.h"
 
 
 /* void	expand_token(t_token *token, char **env)
