--- libft/free_string.c
@@ -1,27 +1,0 @@
-/* ************************************************************************** */
-/*                                                                            */
-/*                                                        :::      ::::::::   */
-/*   free_string.c                                      :+:      :+:    :+:   */
-/*                                                    +:+ +:+         +:+     */
-/*   By: astefane <astefane@student.42madrid.com    +#+  +:+       +#+        */
-/*                                                +#+#+#+#+#+   +#+           */
-/*   Created: 2025/05/03 12:16:35 by astefane          #+#    #+#             */
-/*   Updated: 2025/05/06 16:08:39 by astefane         ###   ########.fr       */
-/*                                                                            */
-/* ************************************************************************** */
-
-#include "libft.h"
-
-void	free_and_error(char *str, char *message, int exit_code, int std)
-{
-	free(str);
-	ft_putstr(message, std);
-	exit(exit_code);
-}
-
-void	free_stuct(t_pipex *data)
-{
-	if (data->pid)
-		free(data->pid);
-	free(data);
-}
