+++ main.c
@@ -6,7 +6,7 @@
 /*   By: astefane <astefane@student.42madrid.com    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/04/12 17:37:01 by astefane          #+#    #+#             */
-/*   Updated: 2025/05/06 18:02:17 by astefane         ###   ########.fr       */
+/*   Updated: 2025/05/10 18:01:20 by astefane         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -32,7 +32,7 @@ static void	process_input(char *input, char **env)
 		current = current->next;
 	}
 	cmd = parse_single_command(list);
-	check_type(list, env);
+	pipex(list, env);
 	free_commands(cmd);
 	free_tokens(list);
 }
